Rename `repository::GitRepository` to `repository::git::Repository`
This follows the following Rust guideline:

https://rust-lang.github.io/rfcs/0356-no-module-prefixes.html
