+++ src/repository.rs
@@ -4,6 +4,3 @@ pub mod signature;
 
 #[cfg(feature = "git")]
 pub mod git;
-
-#[cfg(feature = "git")]
-pub use self::git::GitRepository;
